add config file system

+++ package/sysdef-src/connections.ts
@@ -17,9 +17,7 @@ export const dryFilesystem: Filesystem = {
   async exists(filepath) {
     return fs.existsSync(filepath);
   },
-  async ensureSymlink(destination, source) {
-    console.log(`Creating symlink ${source} -> ${destination}`);
-  },
+  async ensureSymlink(destination, source) {},
   async copy(source, destination) {
     console.log(`Would be copying ${source} -> ${destination}`);
   },
@@ -42,10 +40,9 @@ export const normalFilesystem: Filesystem = {
   },
   async ensureSymlink(destination, source) {
     // Remove existing file or symlink if it exists
-    console.log(`Creating symlink ${source} -> ${destination}`);
-    try {
+    if (fs.existsSync(destination)) {
       const stats = fs.lstatSync(destination);
-      console.log("The file does exist");
+
       if (stats.isFile()) {
         fs.rmSync(destination);
       } else if (stats.isSymbolicLink()) {
@@ -53,12 +50,14 @@ export const normalFilesystem: Filesystem = {
       } else {
         errorOut(`Neither file nor link present in ${destination}. Something is probably not right so we are just stopping`);
       }
-    } catch (error) {
-      // File doesn't exist, which is fine
-      console.log("The file does not exist");
     }
 
-    fs.symlinkSync(source, destination);
+    try {
+      fs.symlinkSync(source, destination);
+    } catch (e) {
+      console.log(e);
+      errorOut(`Error trying to link ${source} -> ${destination}`);
+    }
   },
   async copy(source, destination) {
     const dirname = path.dirname(destination);
